[Notification-Service] rename billablestatus to timecardengine
Named for the domain (time-card-related consumption), not the one event
it happens to handle today, since more time-card-related work is
expected in this consumer group later. Renames the package, its consumer
group env vars (BILLABLE_STATUS_CONSUMER_GROUP/_COUNT ->
TIME_CARD_CONSUMER_GROUP/_COUNT), and default group ID
(csm-notification-service-billable-status -> csm-notification-service-time-card).
No functional change -- Engine.Handle still only logs.
