+++ integrations/csm-notification-service/.env.example
@@ -225,7 +225,7 @@ SLA_CONSUMER_COUNT=
 # defaults to 15s. Accepts any time.ParseDuration string (e.g. "30s", "1m").
 SLA_TICK_INTERVAL=
 
-# Billable status engine — internal/billablestatus, always started (no
+# Billable status engine — internal/timecardengine, always started (no
 # Redis/state dependency, unlike the SLA timer engine above). Consumes
 # case.billable_status_changed on its own dedicated consumer group, same
 # reasoning as SLA_CONSUMER_GROUP above. Currently log-only: entity-service
@@ -234,10 +234,10 @@ SLA_TICK_INTERVAL=
 # this event is itself still commented out — this consumer group exists
 # ahead of need.
 # Consumer group ID this engine's own consumer instances join. Optional —
-# defaults to "csm-notification-service-billable-status".
-BILLABLE_STATUS_CONSUMER_GROUP=
+# defaults to "csm-notification-service-time-card".
+TIME_CARD_CONSUMER_GROUP=
 # How many concurrent consumer instances to run. Optional — defaults to 1.
-BILLABLE_STATUS_CONSUMER_COUNT=
+TIME_CARD_CONSUMER_COUNT=
 
 # Server
 PORT=8080
