+++ integrations/csm-notification-service/internal/timecardengine/engine.go
@@ -14,10 +14,14 @@
 // specific language governing permissions and limitations
 // under the License.
 
-// Package billablestatus reacts to events.TypeCaseBillableStatusChanged —
-// see that type's own doc comment in internal/events/events.go for the full
-// context (entity-service's Postgres data source only, published when a
-// case's severity crosses into or out of LOW).
+// Package timecardengine is the consumer group for time-card-related
+// reactions to case-events records — named for the domain, not the one
+// event it happens to handle today, since more time-card-related work is
+// expected here later. Currently reacts only to
+// events.TypeCaseBillableStatusChanged — see that type's own doc comment in
+// internal/events/events.go for the full context (entity-service's Postgres
+// data source only, published when a case's severity crosses into or out
+// of LOW).
 //
 // Not yet a real implementation: entity-service has no time_cards table/
 // repo/service on its Postgres data source yet — the prerequisite for
@@ -25,7 +29,7 @@
 // Engine.Handle below only logs the event it would eventually act on.
 //
 // This package (and its own dedicated consumer group — see
-// cmd/server/main.go's BILLABLE_STATUS_CONSUMER_GROUP/_COUNT) exists ahead
+// cmd/server/main.go's TIME_CARD_CONSUMER_GROUP/_COUNT) exists ahead
 // of that reaction being buildable, deliberately: the consumer group itself
 // — topic wiring, retry/DLQ behavior, schema validation via internal/events
 // — has no dependency on time_cards at all, only the eventual bulk-update
@@ -36,7 +40,7 @@
 // future bulk update over "several time cards," each its own HTTP round
 // trip to entity-service, must not share a consumer group with
 // latency-sensitive email/Chat dispatch.
-package billablestatus
+package timecardengine
 
 import (
 	"context"
@@ -65,18 +69,18 @@ func NewEngine() *Engine {
 func (e *Engine) Handle(ctx context.Context, record eventbus.Record) error {
 	var env events.Envelope
 	if err := json.Unmarshal(record.Value, &env); err != nil {
-		return fmt.Errorf("billablestatus: decode envelope: %w", err)
+		return fmt.Errorf("timecardengine: decode envelope: %w", err)
 	}
 	if env.Type != events.TypeCaseBillableStatusChanged {
 		return nil
 	}
 	if err := events.Validate(env.EntityID, env.Type, env.Payload); err != nil {
-		return fmt.Errorf("billablestatus: invalid payload: %w", err)
+		return fmt.Errorf("timecardengine: invalid payload: %w", err)
 	}
 
 	var p events.CaseBillableStatusChangedPayload
 	if err := json.Unmarshal(env.Payload, &p); err != nil {
-		return fmt.Errorf("billablestatus: decode payload: %w", err)
+		return fmt.Errorf("timecardengine: decode payload: %w", err)
 	}
 
 	// TODO: bulk-flip every time card's IsBillable for p.CaseID to
@@ -89,6 +93,6 @@ func (e *Engine) Handle(ctx context.Context, record eventbus.Record) error {
 	// call for this event is itself still commented out, so this never
 	// actually receives one in production yet either — this is deliberately
 	// ahead-of-need plumbing, not a live reaction.
-	slog.InfoContext(ctx, "billablestatus: case billable status changed, no consumer action implemented yet", "caseId", p.CaseID, "isBillable", p.IsBillable)
+	slog.InfoContext(ctx, "timecardengine: case billable status changed, no consumer action implemented yet", "caseId", p.CaseID, "isBillable", p.IsBillable)
 	return nil
 }
